cast to float

+++ 0x10-variadic_functions/3-print_all.c
@@ -46,7 +46,7 @@ void printi(va_list valist)
  */
 void printflo(va_list valist)
 {
-	printf("%f", va_arg(valist, double));
+	printf("%f", (float)va_arg(valist, double));
 }
